pyproject.toml
[tool.poetry]
name = "wk-irs"
version = "0.1.0"
description = ""
authors = ["thanawinboon <pokepat27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.1"
celery = "^5.3.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
